refactor(intellij): split CoreProcess into supervisor, RPC channel, and feature routers (#1103)
CoreProcess mixed four responsibilities (process supervision, RPC transport,
message routing, port adapters) in 956 lines. Extract them into a new flat
package `io.miragon.intellij.bpmn.bridge`, mirroring the TS-side `bridge.ts`
decomposition into `composition/`. CoreProcess stays a project `@Service`
(unchanged FQN) acting as façade + composition root.

- RpcHandlerRegistry: string-keyed handler map replacing the monolithic when()
- RpcChannel: NDJSON framing, coalescing outbound queue, writer thread, pumps
- ProcessSupervisor: spawn/backoff-respawn/shutdown-hook/dispose
- BridgeBinaryResolver: binary resolve/extract/cache-prune
- BridgeDeps + six feature routers (editor, deployment, diff, script,
  secret store, host UI), each populating the registry in register()

No behavior change. The former single writeLock splits into RpcChannel.ioLock
(writer + writes) and ProcessSupervisor.stateLock (process identity, held across
attach/detach) to preserve handleExit's check-and-act atomicity; lock order is
always stateLock -> ioLock. dispose() now closes the writer just before removing
the shutdown hook (was just after) as a benign side effect of channel.close().
Lazy-init timing of the port adapters is preserved inside the owning routers.
